Update tgbotapi to v3.

+++ chat_config_repository.go
@@ -20,34 +20,34 @@ func newChatConfigRepository(prefix string, redis *redis.Client) *ChatConfigRepo
 }
 
 // Set - stores any config for chatID
-func (chatConfig *ChatConfigRepository) Set(chatID int, JSON string) {
+func (chatConfig *ChatConfigRepository) Set(chatID int64, JSON string) {
 	chatConfig.redis.Set(chatConfig.ketFor(chatID), JSON, 0)
 }
 
 // SetWithStruct - stores any config for chatID using a struct
-func (chatConfig *ChatConfigRepository) SetWithStruct(chatID int, obj interface{}) {
+func (chatConfig *ChatConfigRepository) SetWithStruct(chatID int64, obj interface{}) {
 	valueBytes, _ := json.Marshal(obj)
 	valueString := string(valueBytes)
 	chatConfig.Set(chatID, valueString)
 }
 
 // Remove - removes config for chatID
-func (chatConfig *ChatConfigRepository) Remove(chatID int) {
+func (chatConfig *ChatConfigRepository) Remove(chatID int64) {
 	chatConfig.redis.Del(chatConfig.ketFor(chatID))
 }
 
 // Get - returns config for chatID
-func (chatConfig *ChatConfigRepository) Get(chatID int) string {
+func (chatConfig *ChatConfigRepository) Get(chatID int64) string {
 	json, _ := chatConfig.redis.Get(chatConfig.ketFor(chatID)).Result()
 	return json
 }
 
 // GetWithStruct - returns config for chatID using a struct
-func (chatConfig *ChatConfigRepository) GetWithStruct(chatID int, obj interface{}) {
+func (chatConfig *ChatConfigRepository) GetWithStruct(chatID int64, obj interface{}) {
 	valueString := chatConfig.Get(chatID)
 	json.Unmarshal([]byte(valueString), obj)
 }
 
-func (chatConfig *ChatConfigRepository) ketFor(chatID int) string {
+func (chatConfig *ChatConfigRepository) ketFor(chatID int64) string {
 	return fmt.Sprintf("%s_%d", chatConfig.key, chatID)
 }
